Change Dynamic method

+++ src/main/java/ua/example/algorithm/FibonacciCalculator.java
@@ -3,6 +3,6 @@ package ua.example.algorithm;
 public interface FibonacciCalculator {
     long findNumberByIndex(int index);
 
-    int INITIAL_FIRST_VALUE = 0;
-    int INITIAL_SECOND_VALUE = 1;
+    Long INITIAL_FIRST_VALUE = 0L;
+    Long INITIAL_SECOND_VALUE = 1L;
 }
